Move Dashboard.tsx to Dashboard folder in components
